// application/src/app/Interfaces/HouseDetails.ts
export interface HouseDetails {
    id:               number;
    title:            string;
    landSize:         number;
    constructionSize: number;
    lat:              number;
    lng:              number;
    kitchenAndLiving: boolean;
    price:            number;
    bedrooms:         number;
    hasPool:          boolean;
    details:          string;
    photo1:           string;
    photo2:           null;
    photo3:           null;
    photo4:           null;
    photo5:           null;
    photo6:           null;
    photo7:           null;
    photo8:           null;
    city:             string;
    parkings:         number;
    sold:             boolean;
    userId:           number;
    mail:             string;
    phone:            string;
    name:             string;
    bathrooms:        number;
}